Use StringView instead of excessive caching in NSString
