+++ Source/SourceKittenFramework/Clang+SourceKitten.swift
@@ -108,7 +108,7 @@ extension CXCursor {
 
     func str() -> String? {
         let cursorExtent = extent()
-        let contents = try! String(contentsOfFile: cursorExtent.start.file, encoding: .utf8)
+        let contents = StringView(try! String(contentsOfFile: cursorExtent.start.file, encoding: .utf8))
         return contents.substringWithSourceRange(start: cursorExtent.start, end: cursorExtent.end)
     }
